+++ api/apiv3/api_match_controller.py
@@ -2,8 +2,8 @@ import json
 
 from google.appengine.ext import ndb
 
-from controllers.apiv3.api_base_controller import ApiBaseController
-from controllers.apiv3.model_properties import filter_match_properties
+from api.apiv3.api_base_controller import ApiBaseController
+from api.apiv3.model_properties import filter_match_properties
 from database.match_query import MatchQuery
