+++ api/apiv3/api_media_controller.py
@@ -1,6 +1,6 @@
 import json
 
-from controllers.apiv3.api_base_controller import ApiBaseController
+from api.apiv3.api_base_controller import ApiBaseController
 from consts.media_tag import MediaTag
