+++ api/apiv3/api_status_controller.py
@@ -1,6 +1,6 @@
 import json
 
-from controllers.apiv3.api_base_controller import ApiBaseController
+from api.apiv3.api_base_controller import ApiBaseController
 
 from models.sitevar import Sitevar
