+++ api/apiv3/api_suggest_controller.py
@@ -1,6 +1,6 @@
 import json
 
-from controllers.apiv3.api_base_controller import ApiBaseController
+from api.apiv3.api_base_controller import ApiBaseController
 from database.team_query import TeamParticipationQuery
 from helpers.suggestions.suggestion_creator import SuggestionCreator
