+++ api/apiv3/model_properties.py
